refactor : CreateChampionshipMatchPanel 불필요한 함수 생성된 부분 제거 했습니다 request status이용하도록 수정했습니다
