+++ src/1_Page/Championship/ui/DashBoard/ui/ChampionshipMatchCardContainer/ui/CreateChampionMatchPanel/model/type.d.ts
@@ -3,3 +3,13 @@ type UseTeamListHandlerReturn = {
   handleAddTeam: (team: ChampionshipTeamInfo) => void;
   handleRemoveTeam: (team: ChampionshipTeamInfo) => void;
 };
+
+type UseManageCreateChampionshipMatchProps = {
+  handleToggleModal: () => void;
+};
+
+type UsePostCreateChampionshipMatchHandlerReturn = {
+  handlePostCreateChampionshipMatch: (
+    formData: CreateChampionMatchFormValues
+  ) => Promise<void>;
+};
