--- src/1_Page/Championship/ui/DashBoard/ui/ChampionshipMatchCardContainer/ui/CreateChampionMatchPanel/util/convert.ts
@@ -1,9 +1,0 @@
-export const convertCreateChampionMatchForm = (
-  data: CreateChampionMatchFormValues
-): UsePostCreateChampionshipMatchProps => {
-  return {
-    first_team_idx: data.teams[0],
-    second_team_idx: data.teams[1],
-    match_match_start_time: `${data.matchDate} ${data.startTime}:00`,
-  };
-};
