+++ src/3_Entity/Championship/usePostCreateChampionshipMatch.ts
@@ -6,16 +6,16 @@ const usePostCreateChampionshipMatch = (
 ): [
   postCreateChampionshipMatch: (
     championshipMatchForm: UsePostCreateChampionshipMatchProps
-  ) => void,
+  ) => Promise<number | undefined>,
   serverState: Record<string, unknown> | null
 ] => {
   const [serverState, request] = useFetchData();
 
-  const postCreateChampionshipMatch = (
+  const postCreateChampionshipMatch = async (
     championshipMatchForm: UsePostCreateChampionshipMatchProps
   ) => {
     const endPoint = `/championship/${championshipListIdx}/championship_match`;
-    request("POST", endPoint, championshipMatchForm, true);
+    return await request("POST", endPoint, championshipMatchForm, true);
   };
 
   React.useEffect(() => {
